+++ src/main/java/com/isep/acme/controllers/ProductController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import com.isep.acme.model.Product;
-import com.isep.acme.model.ProductDTO;
+import com.isep.acme.dtos.ProductDTO;
 
 import com.isep.acme.services.ProductService;
