+++ src/main/java/com/isep/acme/controllers/ReviewController.java
@@ -1,5 +1,8 @@
 package com.isep.acme.controllers;
 
+import com.isep.acme.dtos.CreateReviewDTO;
+import com.isep.acme.dtos.ReviewDTO;
+import com.isep.acme.dtos.VoteReviewDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,7 +10,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
-import com.isep.acme.model.*;
 import com.isep.acme.services.ReviewService;
 
 import java.util.List;
