+++ src/main/java/com/isep/acme/dtos/ImageDTO.java
@@ -1,4 +1,4 @@
-package com.isep.acme.model;
+package com.isep.acme.dtos;
 
 import com.isep.acme.services.ImageService;
