+++ src/main/java/com/isep/acme/dtos/ProductDetailDTO.java
@@ -1,4 +1,4 @@
-package com.isep.acme.model;
+package com.isep.acme.dtos;
 
 public class ProductDetailDTO {
     private String sku;
