+++ src/main/java/com/isep/acme/dtos/ReviewDTO.java
@@ -1,4 +1,4 @@
-package com.isep.acme.model;
+package com.isep.acme.dtos;
 
 import java.time.LocalDate;
