+++ src/main/java/com/isep/acme/model/Product.java
@@ -1,5 +1,7 @@
 package com.isep.acme.model;
 
+import com.isep.acme.dtos.ProductDTO;
+
 import javax.persistence.*;
 import java.util.Objects;
