+++ src/main/java/com/isep/acme/services/ImageService.java
@@ -3,7 +3,7 @@ package com.isep.acme.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 
-import com.isep.acme.model.ImageDTO;
+import com.isep.acme.dtos.ImageDTO;
 import com.isep.acme.model.ProdImage;
 import com.isep.acme.repositories.ImageRepository;
