+++ src/main/java/com/isep/acme/services/ProductService.java
@@ -3,8 +3,8 @@ package com.isep.acme.services;
 import java.util.Optional;
 
 import com.isep.acme.model.Product;
-import com.isep.acme.model.ProductDTO;
-import com.isep.acme.model.ProductDetailDTO;
+import com.isep.acme.dtos.ProductDTO;
+import com.isep.acme.dtos.ProductDetailDTO;
 
 public interface ProductService {
